add roles and some endpoints

+++ API/Controllers/EmployeeController.cs
@@ -6,8 +6,9 @@ using Application.DTOs.Request.Employee;
 using Application.Contracts;
 using Domain.Entities;
 using Application.DTOs.Request.User;
+using System.Threading.Tasks;
 
-[Authorize]
+
 [Route("api/[controller]")]
 [ApiController]
 public class EmployeeController : ControllerBase
@@ -21,19 +22,39 @@ public class EmployeeController : ControllerBase
 
     //Como administrador, quiero ver la lista de empleados registrados para gestionar su información.
     //ya funciona
-    [HttpGet]
-    [AllowAnonymous]
-    // [Authorize(Roles = "Admin")]
+    [HttpGet("GetEmployee")]
     public  List<UsuarioPerfilDto> GetEmployees(string? filtro = null)
     {
+
+        string[] rols = ["Admin"];
+        var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+        if (!a.IsValidUser)
+        {
+            throw new Exception("Usuario no tiene permisos para realiza esta accion");
+        }
         var employees =  _service.GetAllEmployeeAsync(filtro);
+        return employees;
+    }
+
+    [HttpGet("GetMiPerfil")]
+    public async Task<UsuarioPerfilDto> GetMyPerfil()
+    {
+        string[] rols = {
+                "Admin", "DoctorSustituto", "DoctorInterino", "DoctorTitular",
+                "AuxEnfermeria","ATS", "ATSZona","Celadores"};
+
+        var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+        if (!a.IsValidUser)
+        {
+            throw new Exception("Usuario no tiene permisos para realiza esta accion");
+        }
+        var employees = await _service.GetMyPerfilasync(a.UserId);
         return employees;
     }
 
     //se busca la lista de roles
     //ya funciona
     [HttpGet("GetTipoEmployee")]
-    [AllowAnonymous]
     public IActionResult GetRoles()
     {
         var roles = _service.GetRoles();
@@ -46,10 +67,15 @@ public class EmployeeController : ControllerBase
     //la realidad es que se crea el empleado, y el usuario con el rol correspondiente automaticamente.
     //ya funciona
     [HttpPost("register-employee")]
-    [AllowAnonymous]
-    // [Authorize(Roles = "Admin")]
     public async Task<IActionResult> RegisterEmployee([FromBody] RegisterEmployeeDto dto)
     {
+        string[] rols = ["Admin"];
+        var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+        if (!a.IsValidUser)
+        {
+            return Unauthorized("Usuario no tiene permisos para realiza esta accion");
+        }
+        dto.UsuarioCreacion = a.UserId;
         var result = await _service.RegisterUserEmployeAsync(dto);
 
         if (result.Contains("Error"))
@@ -57,26 +83,23 @@ public class EmployeeController : ControllerBase
 
         return Ok(result);
     }
-    //Busca todos los empleados de tipo doctores
-    //ya funciona
-    [HttpGet("Doctores/GetAll")]
-    [AllowAnonymous]
-    // [Authorize(Roles = "Admin")]
-    public async Task<IActionResult> GetAllDoctoresAsync()
-    {
-        var result = await _service.GetAllDoctoresAsync();
-        return Ok(result);
-    }
+
+
     //Como administrador, quiero poder editar la información de un empleado
     //para mantener la información al día.
     //ya funciona
-    [HttpPut]
-    [AllowAnonymous]
-    // [Authorize(Roles = "Admin")]
+    [HttpPut("UpdateUser")]
     public async Task<IActionResult> UpdateEmpleado([FromBody] UpdateEmployeeDto dto)
     {
         try
         {
+            string[] rols = ["Admin"];
+            var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+            if (!a.IsValidUser)
+            {
+                return Unauthorized("Usuario no tiene permisos para realiza esta accion");
+            }
+            dto.UsuarioModificacion = a.UserId;
             var resultado = await _service.UpdateEmpleadoAsync(dto);
             if (!resultado) return NotFound("No se encontró el empleado.");
 
@@ -93,13 +116,18 @@ public class EmployeeController : ControllerBase
     //si ya no trabaja en el centro de salud.
     // tambien elimina el user
     //ya funciona
-    [HttpDelete]
-    [AllowAnonymous]
-    // [Authorize(Roles = "Admin")
+    [HttpDelete("DeleteEmployee")]
     public async Task<IActionResult> DeleteEmpleado([FromBody] DeleteEmployeeDTO deleteEmployee)
     {
         try
         {
+            string[] rols = ["Admin"];
+            var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+            if (!a.IsValidUser)
+            {
+                return Unauthorized("Usuario no tiene permisos para realiza esta accion");
+            }
+            deleteEmployee.IdUserEmployee = a.UserId;
             var resultado = await _service.DeleteEmpleadoAsync(deleteEmployee);
             if (!resultado) return NotFound("No se encontró el empleado.");
 
@@ -113,12 +141,16 @@ public class EmployeeController : ControllerBase
       
     }
     [HttpPut("ActivarEmployee")]
-    [AllowAnonymous]
-    // [Authorize(Roles = "Admin")
     public async Task<IActionResult> ActivarEmpleado(int IdEmployee)
     {
         try
         {
+            string[] rols = ["Admin"];
+            var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+            if (!a.IsValidUser)
+            {
+                return Unauthorized("Usuario no tiene permisos para realiza esta accion");
+            }
             var resultado = await _service.ActivarEmpleadoAsync(IdEmployee);
             if (!resultado) return NotFound("No se encontró el empleado.");
 
@@ -131,5 +163,54 @@ public class EmployeeController : ControllerBase
         }
 
     }
+    //Busca todos los empleados de tipo doctores sustitutos
+    [HttpGet("DoctoresSustitutos/GetAll")]
+    public async Task<IActionResult> GetAllDoctoresSutitutos()
+    {
+        try
+        {
+            string[] rols = ["Admin"];
+            var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+            if (!a.IsValidUser)
+            {
+                return Unauthorized("Usuario no tiene permisos para realiza esta accion");
+            }
+            var resultado = await _service.GetAllDoctoresSustitutosAsync();
+            if (resultado.Count() == 0) return NotFound("No se encontró el empleado.");
+
+            return Ok(resultado);
+        }
+        catch (Exception ex)
+        {
+
+            return BadRequest(ex.Message);
+        }
+
+    }
+    //Busca todos los empleados de tipo doctores no sustitutos
+    [HttpGet("DoctoresNoSustitutos/GetAll")]
+    public async Task<IActionResult> GetAllNoSustituteDoctor()
+    {
+        try
+        {
+            string[] rols = ["Admin"];
+            var a = ValidateToken.validate(Request.Headers["Authorization"].ToString(), rols);
+            if (!a.IsValidUser)
+            {
+                return Unauthorized("Usuario no tiene permisos para realiza esta accion");
+            }
+            var resultado = await _service.GetAllNoSustituteDoctor();
+            if (resultado.Count() == 0) return NotFound("No se encontró el empleado.");
+
+            return Ok(resultado);
+        }
+        catch (Exception ex)
+        {
+
+            return BadRequest(ex.Message);
+        }
+
+    }
+ 
 
 }
